+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.5.0 [01/29/2020]
+### Added
+- Started cone math.
+- Conversions between degrees and radians.
+- dSin(), dCos(), and dAtan2() which works with degrees. 
+
 ## 3.4.0 [08/13/2019]
 ### Added
 - Added Trapezoid math.